Adding valid ranges for Virtualbox
If valid ranges aren't manually set to include 10.100.1.3 the step `potman startvms` will fail with the error:

```
The IP address configured for the host-only network is not within the
allowed ranges. Please update the address used to be within the allowed
ranges and run the command again.

  Address: 10.100.1.3
  Ranges: 192.168.56.0/21

Valid ranges can be modified in the /etc/vbox/networks.conf file. For
more information including valid format see:

  https://www.virtualbox.org/manual/ch06.html#network_hostonly
```

This is fixed by creating `/etc/vbox` and adding the file `/etc/vbox/networks.conf` with contents

```
* 10.100.0.0/16
```